Add files via upload
Added file download_images_combined which has all the code to check if a hotel has a match then downloads it to the server. Also added nogooglemergingJupyter which is the code to check if a hotel has a match in the traffickCam dataset. Also added downloading_images which has exclusively the code to download images to the server.